+++ packages/Compute2/src/components/RFV/RichFunctionView.tsx
@@ -287,6 +287,10 @@ export const RichFunctionView = Vue.defineComponent({
     viewersHook: {
       type: Function as Vue.PropType<ViewersHook>,
     },
+    // standalone hosts that support URL inputs pass a handler; adds a "Copy link with inputs" export
+    urlExportHandler: {
+      type: Function as Vue.PropType<() => void>,
+    },
     view: {
       type: DG.View,
       required: true,
@@ -511,6 +515,8 @@ export const RichFunctionView = Vue.defineComponent({
       activeExports.push(...customExports.value.filter(x => x.function && x.name).map(x => ({...x, handler: () => reportHandler(x.function)})));
       if (canSaveProject())
         activeExports.push({name: 'Save as project...', handler: () => saveCallToProject(currentCall.value, collectDfExportEntries())});
+      if (props.urlExportHandler)
+        activeExports.push({name: 'Copy link with inputs', handler: () => props.urlExportHandler!()});
       return activeExports;
     });