+++ packages/Compute2/src/package-test.ts
@@ -7,6 +7,7 @@ import './test/rfv';
 import './test/navigation';
 import './test/float-display';
 import './test/custom-export';
+import './test/url-inputs';
 import './test/run-comparison-extraction';
 import './test/run-comparison-matching';
 import './test/run-comparison-selection';